requirements.txt
absl-py==2.1.0
asttokens==3.0.0
decorator==5.2.1
executing==2.2.0
immutabledict==4.2.1
ipdb==0.13.13
jedi==0.19.2
matplotlib-inline==0.1.7
networkx==3.4.2
numpy==2.2.3
ortools==9.12.4544
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
protobuf==5.29.3
ptyprocess==0.7.0
pygments==2.19.1
python-dateutil==2.9.0.post0
pytz==2025.1
ruff==0.9.10
six==1.17.0
stack-data==0.6.3
traitlets==5.14.3
tzdata==2025.1
watchdog==6.0.0
wcwidth==0.2.13
